# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Server pets",
        "title": "Swagger Example API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/monitorServer/createServer": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MonitorServer"
                ],
                "summary": "添加新的服务器",
                "parameters": [
                    {
                        "description": "公网IP，私网IP，IPV6，ssh端口，主机名，分组id，状态服务器类型，系统",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MonitorServer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"添加成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/monitorServer/deleteServer/{id}/": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MonitorServer"
                ],
                "summary": "删除服务器",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"删除成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/monitorServer/getServerInfo/{id}/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MonitorServer"
                ],
                "summary": "获取服务器信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/monitorServer/getServerList/{page}/{pageSize}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MonitorServer"
                ],
                "summary": "分页获取服务器列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/monitorServer/putServerInfo/{id}/": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MonitorServer"
                ],
                "summary": "设置服务器",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "公网IP，私网IP，IPV6，ssh端口，主机名，分组id，状态服务器类型，系统",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MonitorServer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"设置成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/system/getLocalServerInfo": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "获取服务器信息",
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/system/getSystemConfig": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "获取配置文件内容",
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/system/reloadSystem": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "重启系统",
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"data\":{},\"msg\":\"重启系统成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/system/setSystemConfig": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "设置配置文件内容",
                "parameters": [
                    {
                        "description": "设置配置文件内容",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.System"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"设置成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "config.AliyunOSS": {
            "type": "object",
            "properties": {
                "accessKeyId": {
                    "type": "string"
                },
                "accessKeySecret": {
                    "type": "string"
                },
                "bucketName": {
                    "type": "string"
                },
                "bucketUrl": {
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                }
            }
        },
        "config.Autocode": {
            "type": "object",
            "properties": {
                "root": {
                    "type": "string"
                },
                "server": {
                    "type": "string"
                },
                "serverApi": {
                    "type": "string"
                },
                "serverInitialize": {
                    "type": "string"
                },
                "serverModel": {
                    "type": "string"
                },
                "serverRequest": {
                    "type": "string"
                },
                "serverRouter": {
                    "type": "string"
                },
                "serverService": {
                    "type": "string"
                },
                "web": {
                    "type": "string"
                },
                "webApi": {
                    "type": "string"
                },
                "webFlow": {
                    "type": "string"
                },
                "webForm": {
                    "type": "string"
                },
                "webTable": {
                    "type": "string"
                }
            }
        },
        "config.Captcha": {
            "type": "object",
            "properties": {
                "imgHeight": {
                    "description": "图片高度",
                    "type": "integer"
                },
                "imgWidth": {
                    "description": "图片宽度",
                    "type": "integer"
                },
                "keyLong": {
                    "description": "验证码长度",
                    "type": "integer"
                }
            }
        },
        "config.Casbin": {
            "type": "object",
            "properties": {
                "modelPath": {
                    "description": "Model路径",
                    "type": "string"
                }
            }
        },
        "config.Detail": {
            "type": "object",
            "properties": {
                "compareField": {
                    "type": "string"
                },
                "interval": {
                    "type": "string"
                },
                "tableName": {
                    "type": "string"
                }
            }
        },
        "config.Email": {
            "type": "object",
            "properties": {
                "from": {
                    "description": "收件人",
                    "type": "string"
                },
                "host": {
                    "description": "服务器地址",
                    "type": "string"
                },
                "isSSL": {
                    "description": "是否SSL",
                    "type": "boolean"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "port": {
                    "description": "端口",
                    "type": "integer"
                },
                "secret": {
                    "description": "密钥",
                    "type": "string"
                },
                "to": {
                    "description": "收件人",
                    "type": "string"
                }
            }
        },
        "config.Excel": {
            "type": "object",
            "properties": {
                "dir": {
                    "type": "string"
                }
            }
        },
        "config.JWT": {
            "type": "object",
            "properties": {
                "bufferTime": {
                    "description": "缓冲时间",
                    "type": "integer"
                },
                "expiresTime": {
                    "description": "过期时间",
                    "type": "integer"
                },
                "signingKey": {
                    "description": "jwt签名",
                    "type": "string"
                }
            }
        },
        "config.Local": {
            "type": "object",
            "properties": {
                "path": {
                    "description": "本地文件路径",
                    "type": "string"
                }
            }
        },
        "config.Mysql": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "string"
                },
                "dbname": {
                    "description": "数据库名",
                    "type": "string"
                },
                "logMode": {
                    "type": "boolean"
                },
                "logZap": {
                    "type": "string"
                },
                "maxIdleConns": {
                    "type": "integer"
                },
                "maxOpenConns": {
                    "type": "integer"
                },
                "password": {
                    "description": "数据库密码",
                    "type": "string"
                },
                "path": {
                    "description": "服务器地址:端口",
                    "type": "string"
                },
                "username": {
                    "description": "数据库用户名",
                    "type": "string"
                }
            }
        },
        "config.Qiniu": {
            "type": "object",
            "properties": {
                "accessKey": {
                    "description": "accessKey",
                    "type": "string"
                },
                "bucket": {
                    "description": "空间名称",
                    "type": "string"
                },
                "imgPath": {
                    "description": "CDN加速域名",
                    "type": "string"
                },
                "secretKey": {
                    "description": "secretKey",
                    "type": "string"
                },
                "useCdnDomains": {
                    "description": "上传是否使用CDN上传加速",
                    "type": "boolean"
                },
                "useHttps": {
                    "description": "是否使用https",
                    "type": "boolean"
                },
                "zone": {
                    "description": "存储区域",
                    "type": "string"
                }
            }
        },
        "config.Redis": {
            "type": "object",
            "properties": {
                "addr": {
                    "description": "服务器地址:端口",
                    "type": "string"
                },
                "db": {
                    "type": "integer"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                }
            }
        },
        "config.Server": {
            "type": "object",
            "properties": {
                "aliyunOSS": {
                    "$ref": "#/definitions/config.AliyunOSS"
                },
                "autoCode": {
                    "description": "auto",
                    "$ref": "#/definitions/config.Autocode"
                },
                "captcha": {
                    "$ref": "#/definitions/config.Captcha"
                },
                "casbin": {
                    "$ref": "#/definitions/config.Casbin"
                },
                "email": {
                    "$ref": "#/definitions/config.Email"
                },
                "excel": {
                    "$ref": "#/definitions/config.Excel"
                },
                "jwt": {
                    "$ref": "#/definitions/config.JWT"
                },
                "local": {
                    "description": "oss",
                    "$ref": "#/definitions/config.Local"
                },
                "mysql": {
                    "description": "gorm",
                    "$ref": "#/definitions/config.Mysql"
                },
                "qiniu": {
                    "$ref": "#/definitions/config.Qiniu"
                },
                "redis": {
                    "$ref": "#/definitions/config.Redis"
                },
                "system": {
                    "$ref": "#/definitions/config.System"
                },
                "tencentCOS": {
                    "$ref": "#/definitions/config.TencentCOS"
                },
                "timer": {
                    "$ref": "#/definitions/config.Timer"
                },
                "zap": {
                    "$ref": "#/definitions/config.Zap"
                }
            }
        },
        "config.System": {
            "type": "object",
            "properties": {
                "addr": {
                    "description": "端口值",
                    "type": "integer"
                },
                "dbType": {
                    "description": "数据库类型:mysql(默认)|sqlite|sqlserver|postgresql",
                    "type": "string"
                },
                "env": {
                    "description": "环境值",
                    "type": "string"
                },
                "ossType": {
                    "description": "Oss类型",
                    "type": "string"
                },
                "useMultipoint": {
                    "description": "多点登录拦截",
                    "type": "boolean"
                }
            }
        },
        "config.TencentCOS": {
            "type": "object",
            "properties": {
                "baseURL": {
                    "type": "string"
                },
                "bucket": {
                    "type": "string"
                },
                "pathPrefix": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "secretID": {
                    "type": "string"
                },
                "secretKey": {
                    "type": "string"
                }
            }
        },
        "config.Timer": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/config.Detail"
                    }
                },
                "spec": {
                    "type": "string"
                },
                "start": {
                    "type": "boolean"
                }
            }
        },
        "config.Zap": {
            "type": "object",
            "properties": {
                "director": {
                    "description": "日志文件夹",
                    "type": "string"
                },
                "encodeLevel": {
                    "description": "编码级",
                    "type": "string"
                },
                "format": {
                    "description": "输出",
                    "type": "string"
                },
                "level": {
                    "description": "级别",
                    "type": "string"
                },
                "linkName": {
                    "description": "软链接名称",
                    "type": "string"
                },
                "logInConsole": {
                    "description": "输出控制台",
                    "type": "boolean"
                },
                "prefix": {
                    "description": "日志前缀",
                    "type": "string"
                },
                "showLine": {
                    "description": "显示行",
                    "type": "boolean"
                },
                "stacktraceKey": {
                    "description": "栈名",
                    "type": "string"
                }
            }
        },
...
        "model.MonitorServer": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "groupId": {
                    "type": "string"
                },
                "hostName": {
                    "type": "string"
                },
                "id": {
                    "description": "主键ID",
                    "type": "integer"
                },
                "ipV6Address": {
                    "type": "string"
                },
                "keyFile": {
                    "type": "string"
                },
                "pubIpAddress": {
                    "type": "string"
                },
                "pveIpAddress": {
                    "type": "string"
                },
                "serverType": {
                    "type": "integer"
                },
                "sshPort": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                },
                "system": {
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "model.System": {
            "type": "object",
            "properties": {
                "config": {
                    "$ref": "#/definitions/config.Server"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}